project/requirements.txt
Flask==3.0.0
flask-mysqldb==2.0.0
flask-cors==4.0.0
PyJWT==2.8.0
bcrypt==5.0.0
python-dotenv
twilio==8.1.0
google-generativeai
